began config file parsing
