+++ Cargo.toml
@@ -8,6 +8,7 @@ readme = "README.md"
 repository = "https://github.com/tkatter/sericom"
 keywords = ["serial", "console"]
 categories = ["command-line-utilities"]
+exclude = [".github/*"]
 
 [package.metadata.deb]
 maintainer = "Thomas Katter <tjkatter@gmail.com>"
@@ -18,8 +19,10 @@ license-file = ["LICENSE", "14"]
 chrono = "0.4.41"
 clap = { version = "4.5.41", features = ["derive"] }
 crossterm = { version = "0.29.0", features = ["osc52", "event-stream"] }
+serde = { version = "1.0.219", features = ["derive"] }
 serial2-tokio = { version = "0.1.15", features = ["windows"] }
 tokio = { version = "1.47.0", features = ["full"] }
+toml = "0.9.5"
 # tracing = "0.1"
 # console-subscriber = "0.4.1"
